separate out vsvip rest cache logic
